[REFACTOR/#71] ReviewCard entities 레이어 이동 및 StarIcon 공유 컴포넌트 분리
