+++ src/entities/review/index.ts
@@ -1,0 +1,2 @@
+export type { Review, ReviewImage } from "./model/types";
+export { ReviewCard } from "./ui/ReviewCard";
